requirements.txt
# ===================================================================
# YOLO11 Training Requirements
# Optimized for: RTX 5080 (16GB) + Intel Ultra 7 265K + 64GB RAM
# CUDA Version: 12.8
# Python Version: 3.8-3.11 (3.10 recommended)
# ===================================================================

# NOTE: Install PyTorch FIRST with CUDA 12.8 support:
# python -m pip install --upgrade --force-reinstall torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128

# ===================================================================
# Core Dependencies (Required)
# ===================================================================

# YOLO Framework
ultralytics>=8.0.0              # YOLOv11 framework

# Computer Vision
opencv-python>=4.8.0            # Image/video processing
opencv-contrib-python>=4.8.0    # Additional OpenCV features (optional)
Pillow>=10.0.0                  # Image handling

# Data Processing
numpy>=1.24.0                   # Numerical computing
pandas>=2.0.0                   # Data analysis
PyYAML>=6.0                     # YAML configuration files

# System Monitoring
psutil>=5.9.0                   # System and process utilities
tqdm>=4.65.0                    # Progress bars

# Visualization
matplotlib>=3.7.0               # Plotting
seaborn>=0.12.0                 # Statistical visualizations

# ===================================================================
# Model Export & Deployment (Recommended)
# ===================================================================

# ONNX Export (Universal format - Recommended)
onnx>=1.15.0                    # ONNX model format
onnxruntime-gpu>=1.16.0         # ONNX inference with GPU support
onnxsim>=0.4.0                  # ONNX model simplification

# TensorRT (NVIDIA GPU optimization - Highly Recommended for RTX 5080)
# nvidia-tensorrt>=10.0.0       # Uncomment if using TensorRT export
# Note: TensorRT requires separate installation from NVIDIA

# OpenVINO (Intel CPU optimization)
# openvino>=2024.0.0            # Uncomment for Intel optimization

# ===================================================================
# Training Monitoring & Logging (Recommended)
# ===================================================================

tensorboard>=2.15.0             # Training visualization
# wandb>=0.16.0                 # Weights & Biases (optional cloud logging)
# mlflow>=2.10.0                # MLflow experiment tracking (optional)

# ===================================================================
# Data Augmentation (Optional but Recommended)
# ===================================================================

albumentations>=1.3.0           # Advanced image augmentation
imgaug>=0.4.0                   # Image augmentation library

# ===================================================================
# Utilities (Optional)
# ===================================================================

# Dataset tools
labelImg>=1.8.0                 # Image labeling tool (if labeling data)
# roboflow>=1.1.0               # Roboflow dataset integration (optional)

# Performance profiling
py-cpuinfo>=9.0.0               # CPU information
GPUtil>=1.4.0                   # GPU monitoring

# HTTP/API
requests>=2.31.0                # HTTP library
urllib3>=2.0.0                  # HTTP client

# ===================================================================
# Mobile & Edge Deployment (Optional)
# ===================================================================

# TensorFlow Lite (for mobile deployment)
# tensorflow>=2.15.0            # Uncomment for TFLite export
# Note: Large package, only install if needed

# CoreML (for Apple devices)
# coremltools>=7.0              # macOS/iOS only

# ===================================================================
# Development Tools (Optional)
# ===================================================================

# Code quality
# black>=23.0.0                 # Code formatter
# flake8>=6.0.0                 # Linter
# pytest>=7.4.0                 # Testing framework

# Jupyter notebooks
# jupyter>=1.0.0                # Jupyter notebook
# ipywidgets>=8.0.0             # Interactive widgets

# ===================================================================
# Version Constraints
# ===================================================================

# Ensure compatibility
scipy>=1.10.0                   # Scientific computing
scikit-learn>=1.3.0             # Machine learning utilities
torch>=2.1.0                    # PyTorch (installed separately with CUDA)
torchvision>=0.16.0             # PyTorch vision (installed separately)

# ===================================================================
# Installation Instructions
# ===================================================================

# 1. Create virtual environment:
#    python -m venv venv
#    .\venv\Scripts\Activate.ps1  (Windows PowerShell)
#
# 2. Install PyTorch with CUDA 12.8 (MUST DO FIRST):
#    python -m pip install --upgrade --force-reinstall torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128
#
# 3. Install other requirements:
#    pip install -r requirements.txt
#
# 4. Verify installation:
#    python check_setup.py
#
# ===================================================================
# RTX 5080 Specific Notes
# ===================================================================

# - CUDA 12.8 provides best performance for Blackwell architecture
# - TensorRT 10.0+ recommended for maximum inference speed
# - Ensure NVIDIA driver version 560.0 or higher
# - Use onnxruntime-gpu for ONNX inference acceleration
#
# ===================================================================
